+++ datastore/catalog/memory/db_controller.go
@@ -1,6 +1,7 @@
 package memory
 
 import (
+	"context"
 	"database/sql"
 	"errors"
 
@@ -8,8 +9,8 @@ import (
 )
 
 type DB interface {
-	Query(q string, args ...any) (*sql.Rows, error)
-	Exec(q string, args ...any) (sql.Result, error)
+	QueryContext(ctx context.Context, query string, args ...any) (*sql.Rows, error)
+	ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error)
 }
 
 var _ DB = &dbController{}
@@ -23,33 +24,33 @@ type dbController struct {
 	base *sql.DB
 }
 
-func (d *dbController) Query(q string, args ...any) (*sql.Rows, error) {
+func (d *dbController) QueryContext(ctx context.Context, query string, args ...any) (*sql.Rows, error) {
 	if d.tx != nil {
-		return d.tx.Query(q, args...)
+		return d.tx.QueryContext(ctx, query, args...)
 	}
 
-	return d.base.Query(q, args...)
+	return d.base.QueryContext(ctx, query, args...)
 }
 
-func (d *dbController) Exec(q string, args ...any) (sql.Result, error) {
+func (d *dbController) ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error) {
 	if d.tx != nil {
-		return d.tx.Exec(q, args...)
+		return d.tx.ExecContext(ctx, query, args...)
 	}
 
-	return d.base.Exec(q, args...)
+	return d.base.ExecContext(ctx, query, args...)
 }
 
-// Fixture performs an Exec but ignores the result, and is intended for test setup
-func (d *dbController) Fixture(q string, args ...any) error {
-	_, err := d.Exec(q, args...)
+// Fixture performs an ExecContext but ignores the result, and is intended for test setup
+func (d *dbController) Fixture(ctx context.Context, query string, args ...any) error {
+	_, err := d.ExecContext(ctx, query, args...)
 	return err
 }
 
-func (d *dbController) Begin() error {
+func (d *dbController) Begin(ctx context.Context) error {
 	if d.tx != nil {
 		return errors.New("transaction already started")
 	}
-	tx, err := d.base.Begin()
+	tx, err := d.base.BeginTx(ctx, nil)
 	if err != nil {
 		return err
 	}